Fix items list update on category change / Remove duplicated component
There was and issue in the Items component that prevented the list update when the category has changed. It was caused because the only place where the list change hadler was being called was in componentDidMount. In this case, when the prop got updated, the component re-rendered but it was not re-mounted. To solve id, I have added the componentDidUpdate life cycle method validating if the category got changed and triggered the list items fetch.
I have also removed the Home component as it was a duplicate of the CategoryItems component, with the logic added to the later one to render the rigth thing whenever the component is rendered